Updated type file and renamed constants.js
