+++ client/src/actions/constants.js
@@ -1,0 +1,3 @@
+export const GET_MEAL_INFO = "GET_MEAL_INFO";
+
+export const ADD_TO_BASKET = "ADD_TO_BASKET";
