+++ client/src/actions/getMealInfo.js
@@ -1,11 +1,11 @@
 import axios from "axios";
-import { GET_MEAL_INFO } from "./types";
+import { GET_MEAL_INFO } from "./constants";
 
 const getMealInfo = mealId => {
   return dispatch => {
     dispatch({
       type: GET_MEAL_INFO,
-      payload: axios.get(`/api/mealInfo/${mealId}`),
+      payload: axios.get(`/api/mealInfo/${mealId}`)
     });
   };
 };
