+++ client/src/reducers/addBasketReducer.js
@@ -1,16 +1,16 @@
-import { ADD_TO_BASKET } from '../actions/types';
+import { ADD_TO_BASKET } from "../actions/constants";
 
 const initialState = {
-  basket: [],
+  basket: []
 };
 
-export default function (state = initialState, action) {
-  let {basket : oldBasket} = state;
+export default function(state = initialState, action) {
+  let { basket: oldBasket } = state;
   switch (action.type) {
     case ADD_TO_BASKET:
       return {
         ...state,
-        basket: [...oldBasket,action.payload],
+        basket: [...oldBasket, action.payload]
       };
     default:
       return state;
