+++ client/src/reducers/getMealInfoReducer.js
@@ -1,16 +1,15 @@
-import { GET_MEAL_INFO_PENDING, GET_MEAL_INFO_FULFILLED, GET_MEAL_INFO_REJECTED} from '../actions/types';
+import { GET_MEAL_INFO } from "../actions/constants";
 
-const initialState = {
-};
+const initialState = {};
 
-export default function (state = initialState, action) {
+export default function(state = initialState, action) {
   switch (action.type) {
-    case GET_MEAL_INFO_FULFILLED:
-      return {mealData:action.payload.data, status:"Fulfilled"};
-    case GET_MEAL_INFO_PENDING:
-      return {status:"Pending"};
-    case GET_MEAL_INFO_REJECTED:
-      return {status:"Rejected"}
+    case `${GET_MEAL_INFO}_FULFILLED`:
+      return { mealData: action.payload.data, status: "Fulfilled" };
+    case `${GET_MEAL_INFO}_PENDING`:
+      return { status: "Pending" };
+    case `${GET_MEAL_INFO}_REJECTED`:
+      return { status: "Rejected" };
     default:
       return state;
   }
